video_generator/requirements.txt
openai
python-dotenv
opencv-python
numpy
moviepy<=2.0
gTTS
loguru